+++ R/filtering.R
@@ -304,41 +304,25 @@ filtering <- function(gmdbn, evid, nodes = names(gmdbn$b_1), col_seq = NULL,
 
       n_nodes_gmdbn <- nodes_gmdbn %>%
         length()
-      arcs <- struct$arcs %>%
-        bind_rows()
 
       if (n_nodes < n_nodes_gmdbn) {
-        nodes_obs <- evid %>%
-          select(any_of(nodes_gmdbn) & where(~ !any(is.na(.)))) %>%
+        evid_nodes <- evid %>%
+          select(any_of(nodes_gmdbn))
+        nodes_obs <- evid_nodes %>%
+          select(where(~ !any(is.na(.)))) %>%
           colnames()
-        blanket <- nodes
-        blanket_miss <- blanket[!(blanket %in% nodes_obs)]
-
-        while (length(blanket_miss) > 0) {
-          child <- arcs %>%
-            filter(from %in% blanket_miss) %>%
-            .$to
-          copar <- arcs %>%
-            filter(to %in% c(blanket_miss, child)) %>%
-            .$from
-          blanket_add <- child %>%
-            c(copar) %>%
-            setdiff(blanket)
-          blanket <- blanket %>%
-            c(blanket_add)
-          blanket_miss <- blanket_add[!(blanket_add %in% nodes_obs)]
-        }
-
+        nodes_miss <- evid_nodes %>%
+          select(where(~ all(is.na(.)))) %>%
+          colnames() %>%
+          c(setdiff(nodes_gmdbn, colnames(evid_nodes)))
         gmdbn <- gmdbn %>%
-          remove_nodes(setdiff(nodes_gmdbn, blanket))
+          relevant(nodes, nodes_obs, nodes_miss)
         nodes_gmdbn <- gmdbn$b_1 %>%
           names()
         n_nodes_gmdbn <- nodes_gmdbn %>%
           length()
         col_evid_prop <- col_seq %>%
           c(nodes_gmdbn)
-        arcs <- arcs %>%
-          filter(from %in% blanket, to %in% blanket)
       }
 
       col_sub <- prefix %>%
@@ -348,7 +332,10 @@ filtering <- function(gmdbn, evid, nodes = names(gmdbn$b_1), col_seq = NULL,
       col_prop <- col_seq %>%
         c(col_weight)
       max_lag <- lag[n_lags]
-      n_prop <- arcs$lag %>%
+      n_prop <- struct$arcs %>%
+        bind_rows() %>%
+        filter(from %in% nodes_gmdbn, to %in% nodes_gmdbn) %>%
+        .$lag %>%
         max(max_lag) - 1
 
       if (n_prop >= 0) {