+++ R/remove_arcs.R
@@ -17,7 +17,8 @@
 #' @return The \code{gmbn} or \code{gmdbn} object after removing the arcs.
 #'
 #' @seealso \code{\link{add_arcs}}, \code{\link{add_nodes}},
-#' \code{\link{remove_nodes}}, \code{\link{rename_nodes}}
+#' \code{\link{relevant}}, \code{\link{remove_nodes}},
+#' \code{\link{rename_nodes}}
 #'
 #' @examples
 #' data(gmbn_body)